create authhelper.test file and moved logout tests to that file
