+++ context/AuthContextProvider.test.tsx
@@ -3,7 +3,7 @@ import { account } from '../api/config';
 import Alert from '../components/AlertNotification/AlertNotification';
 import { AlertVariants } from '../components/AlertNotification/Alerts.enum';
 import { toast } from 'react-hot-toast';
-import { loginAccount, logoutHandler } from './AuthHelper';
+import { loginAccount } from './AuthHelper';
 import { AppRouterInstance } from 'next/dist/shared/lib/app-router-context.shared-runtime';
 
 const mockCreateEmailPasswordSession = jest.fn();
@@ -31,9 +31,6 @@ describe('AuthContextProvider', () => {
     email: 'testemail@email.com',
     password: 'password1234',
   });
-
-  const setIsSignedIn = jest.fn();
-  const resetUser = jest.fn();
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -68,28 +65,6 @@ describe('AuthContextProvider', () => {
     expect(error).toEqual(mockError);
     expect(toast.custom).toHaveBeenCalledWith(
       <Alert variant={AlertVariants.Error} message="Something went wrong!" />,
-    );
-  });
-
-  //logout tests
-  test('after a successful logout it shows success notification', async () => {
-    await logoutHandler({ resetUser, setIsSignedIn, router });
-
-    expect(resetUser).toHaveBeenCalled();
-    expect(router.push).toHaveBeenCalledWith('/login');
-    expect(setIsSignedIn).toHaveBeenCalledWith(false);
-    expect(toast.custom).toHaveBeenCalledWith(
-      <Alert variant={AlertVariants.Success} message="Logged Out" />,
-    );
-  });
-
-  test('after logout attempt errors it shows error notification', async () => {
-    account.deleteSession = jest.fn().mockRejectedValue('test error');
-
-    const error = await logoutHandler({ resetUser, setIsSignedIn, router });
-
-    expect(toast.custom).toHaveBeenCalledWith(
-      <Alert variant={AlertVariants.Error} message="Logout failed!" />,
     );
   });
 });
